refactor(vo): 값객체 패키지 이동 domain -> vo
