+++ src/main/java/domain/Lottos.java
@@ -5,6 +5,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import vo.WinningCount;
 
 public class Lottos {
     private final List<Lotto> lottos;
