+++ src/main/java/vo/InputMoney.java
@@ -1,12 +1,12 @@
-package domain;
+package vo;
 
 import java.util.Objects;
 
 public class InputMoney {
-    private static final String ERROR_MESSAGE_FOR_INVALID_INPUT_MONEY = "금액은 1000원 이상, 1000원 단위로 입력해주세요";
     private static final int MINIMUM_MONEY = 1000;
     private static final int INPUT_MONEY_UNIT = 1000;
     private static final int REMAIN = 0;
+    private static final String ERROR_MESSAGE_FOR_INVALID_INPUT_MONEY = "금액은 1000원 이상, 1000원 단위로 입력해주세요";
 
     private final int money;
