+++ src/main/java/vo/TrialNumber.java
@@ -1,4 +1,4 @@
-package domain;
+package vo;
 
 import java.util.Objects;
