+++ src/main/java/vo/WinningCount.java
@@ -1,8 +1,11 @@
-package domain;
+package vo;
 
 import java.util.Objects;
 
 public class WinningCount {
+    private static final int MINIMUM_WINNING_COUNT = 0;
+    private static final String ERROR_MESSAGE_FOR_MINUS_WINNING_COUNT = "[ERROR] 당첨 횟수는 음수일 수 없습니다.";
+
     private final int count;
 
     public int getCount() {
@@ -10,7 +13,14 @@ public class WinningCount {
     }
 
     public WinningCount(int count) {
+        validate(count);
         this.count = count;
+    }
+
+    private void validate(int count) {
+        if (count < MINIMUM_WINNING_COUNT) {
+            throw new IllegalArgumentException(ERROR_MESSAGE_FOR_MINUS_WINNING_COUNT);
+        }
     }
 
     @Override
