+++ src/test/java/domain/LottosTest.java
@@ -12,6 +12,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import vo.WinningCount;
 
 public class LottosTest {
     private List<Lotto> lottos = new ArrayList<>();
