+++ src/test/java/vo/TrialNumberTest.java
@@ -1,4 +1,4 @@
-package domain;
+package vo;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
